refactor(upload): split monolithic main.rs into modules
